Use size_t in filedatasource.

+++ libdf3d/io/FileDataSource.h
@@ -21,10 +21,10 @@ public:
     virtual bool valid() const = 0;
 
     virtual size_t getRaw(void *buffer, size_t sizeInBytes) = 0;
-    virtual int64_t getSizeInBytes() = 0;
+    virtual size_t getSizeInBytes() = 0;
 
-    virtual int64_t tell() = 0;
-    virtual bool seek(int64_t offset, std::ios_base::seekdir origin) = 0;
+    virtual size_t tell() = 0;
+    virtual bool seek(size_t offset, std::ios_base::seekdir origin) = 0;
 
     template<typename T>
     bool getAsObjects(T *output, size_t numElements)
