+++ libdf3d/io/MemoryDataSource.h
@@ -8,19 +8,19 @@ class MemoryDataSource : public FileDataSource
 {
     unique_ptr<unsigned char[]> m_buffer;
     unsigned char *m_current;
-    int64_t m_size;
+    int32_t m_size;
 
 public:
-    MemoryDataSource(unique_ptr<unsigned char[]> &&buffer, int64_t size, const std::string &fileName);
+    MemoryDataSource(unique_ptr<unsigned char[]> &&buffer, int32_t size, const std::string &fileName);
     ~MemoryDataSource();
 
     bool valid() const override;
 
     size_t getRaw(void *buffer, size_t sizeInBytes) override;
-    int64_t getSizeInBytes() override;
+    size_t getSizeInBytes() override;
 
-    int64_t tell() override;
-    bool seek(int64_t offset, std::ios_base::seekdir origin) override;
+    size_t tell() override;
+    bool seek(size_t offset, std::ios_base::seekdir origin) override;
 };
 
 }
