+++ libdf3d/resources/ResourceContainer.h
@@ -48,7 +48,7 @@ class ResourceContainerDataSource : public FileDataSource
 {
     weak_ptr<FileDataSource> m_archiveFile;
     ResourceContainer::Entry m_entry;
-    int64_t m_internalPos = 0;
+    size_t m_internalPos = 0;
 
 public:
     ResourceContainerDataSource(weak_ptr<FileDataSource> archiveFile, const ResourceContainer::Entry &entry);
@@ -57,10 +57,10 @@ public:
     bool valid() const override;
 
     size_t getRaw(void *buffer, size_t sizeInBytes) override;
-    int64_t getSizeInBytes() override;
+    size_t getSizeInBytes() override;
 
-    int64_t tell() override;
-    bool seek(int64_t offset, std::ios_base::seekdir origin) override;
+    size_t tell() override;
+    bool seek(size_t offset, std::ios_base::seekdir origin) override;
 };
 
 }
